requirements.txt
tavily-python>=0.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
# Agent stack: LangChain, LangGraph, Ollama
langchain-core>=0.2.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
pydantic>=2.0.0
chromadb>=0.4.0
# Telegram bot
python-telegram-bot>=21.0
